warning fixed

+++ main.cpp
@@ -51,7 +51,7 @@ int main(int argc, char* argv[])
                       static_cast<double>(label->height());
       int img_h = static_cast<int>(
           ceil(static_cast<double>(label->height()) * img_quality));
-      int img_w = ceil(((double) img_h) * aspect);
+      int img_w = static_cast<int>(ceil(static_cast<double>(img_h) * aspect));
       image = new QImage(img_w, img_h, QImage::Format_ARGB32);
       label->setImage(image);
     }
